+++ agent/crates/l7-parser/src/mysql/parse.rs
@@ -8,15 +8,15 @@ use nom::{
 	number::streaming::{le_u8, le_u24},
 };
 use observ_trace_common::{Direction, MessageType};
-fn payload_length(expected: usize) -> impl Fn(&[u8]) -> IResult<&[u8], u32> {
-	move |i| verify(le_u24, |&payload_length| payload_length as usize + 4 == expected).parse(i)
+fn payload_length(expected: u32) -> impl Fn(&[u8]) -> IResult<&[u8], u32> {
+	move |i| verify(le_u24, |&payload_length| payload_length + 4 == expected).parse(i)
 }
 fn com(i: &[u8]) -> IResult<&[u8], Com> {
 	map_res(le_u8, |com: u8| Com::try_from(com).map_err(|_| Error::new(i, ErrorKind::MapRes)))
 		.parse(i)
 }
 
-pub(super) fn mysql(i: &[u8], count: usize, direction: Direction) -> Result<MySQL> {
+pub(super) fn mysql(i: &[u8], count: u32, direction: Direction) -> Result<MySQL> {
 	let mut header = (payload_length(count), le_u8, com);
 	let (_, (payload_length, _sequence_id, _)) = header.parse(i).map_err(|_| PARSE_MYSQL_FAILED)?;
